Support for additional syntaxes: $num$ and $num,name=var$. Defaults to gaussian distribution when no explicit distribution name given. Includes ITs.

+++ Testing/FileProcessingServiceIT.py
@@ -7,6 +7,7 @@ import unittest
 from FileProcessingService import FileProcessingService
 from Sim1FileProcessingService import Sim1FileProcessingService
 from SupportedFileTypes import SupportedFileTypes
+from SupportedDistributions import SupportedDistributions
 
 
 class FileProcessingServiceIT(unittest.TestCase):
@@ -20,7 +21,7 @@ class FileProcessingServiceIT(unittest.TestCase):
     def tearDown(self):
         if self.file_processing_service:
             self.file_processing_service.data_file.close()
-        if self.generated_folder != "/":
+        if self.generated_folder != "/" and os.path.isdir(self.generated_folder):
             shutil.rmtree(self.generated_folder)
 
     def setupFileProcessingService(self, data_file, file_type):
@@ -38,15 +39,15 @@ class FileProcessingServiceIT(unittest.TestCase):
     def testMATLABGenomeFilesSuccessfullyCreated(self):
         data_file = self.setTargetFile('SampleDataFiles', 'WNT_ERK_crosstalk.m')
         self.setupFileProcessingService(data_file, SupportedFileTypes.MATLAB)
-        self.assertFilesCreatedSuccessfully(SupportedFileTypes.MATLAB)
+        self.assertGenomeFilesCreatedSuccessfully(SupportedFileTypes.MATLAB)
 
     def testRGenomeFilesSuccessfullyCreated(self):
         # Note: This is R the program, not R an integer representing permutations.
         data_file = self.setTargetFile('SampleDataFiles', 'booleanModel.r.t')
         self.setupFileProcessingService(data_file, SupportedFileTypes.R)
-        self.assertFilesCreatedSuccessfully(SupportedFileTypes.R)
+        self.assertGenomeFilesCreatedSuccessfully(SupportedFileTypes.R)
 
-    def assertFilesCreatedSuccessfully(self, file_type):
+    def assertGenomeFilesCreatedSuccessfully(self, file_type):
         number_of_genomes = self.file_processing_service.number_of_genomes
         self.log.info("Testing %s genomes are successfully created as %s files.", number_of_genomes, file_type)
 
@@ -72,13 +73,104 @@ class FileProcessingServiceIT(unittest.TestCase):
         self.log.info("Testing %s genomes are successfully created as MATLAB files.", number_of_genomes)
         self.file_processing_service.createGenomes()
 
+        self.assertSim1TrialFilesSuccessfullyCreated(number_of_genomes, number_of_trials)
+
+    def testSim1FileProcessingServiceForR(self):
+        data_file = self.setTargetFile('SampleDataFiles', 'booleanModel.r.t')
+        self.setupFileProcessingService(data_file, SupportedFileTypes.R)
+
+        data_file_for_sim1 = self.setTargetFile('SampleDataFiles', 'booleanTest.r.u')
+        self.setupFileProcessingServiceForSim1(data_file_for_sim1, SupportedFileTypes.R)
+
+        number_of_genomes = self.sim1FileProcessingService.number_of_genomes
+        number_of_trials = self.sim1FileProcessingService.number_of_trials
+
+        self.log.info("Testing %s genomes are successfully created as R files.", number_of_genomes)
+        self.file_processing_service.createGenomes()
+
+        self.assertSim1TrialFilesSuccessfullyCreated(number_of_genomes, number_of_trials)
+
+    def assertSim1TrialFilesSuccessfullyCreated(self, number_of_genomes, number_of_trials):
         trial_files_created = self.sim1FileProcessingService.createTrialFiles()
-
         assert len(trial_files_created) == number_of_trials * number_of_genomes
-
         assert os.path.isdir(self.current_working_dir)
         total_created_files = [file for file in os.listdir(self.generated_folder)]
-
         # 2*K files created + R*K trial files
         assert len(total_created_files) == (2 * number_of_genomes) + len(trial_files_created)
-        assert len([file for file in total_created_files if re.findall(r'trial\d*_genome\d*.*', file)]) == len(trial_files_created)
+        assert len([file for file in total_created_files if re.findall(r'trial\d*_genome\d*.*', file)]) == len(
+            trial_files_created)
+
+    def testDollarSignSyntaxValuesCorrectlyExtractedForR(self):
+        data_file = self.setTargetFile('SampleDataFiles', 'booleanModel.r.t')
+        self.setupFileProcessingService(data_file, SupportedFileTypes.R)
+        fp_service = self.file_processing_service
+
+        line_default = 'variableOne <- $gauss(0,1),name=varOne$;'
+        target_sequence_default = self.extractTargetSequenceFromLine(line_default)
+        assert fp_service.extractCoefficientName(target_sequence_default) == 'varOne'
+        assert fp_service.extractDistributionName(target_sequence_default) == SupportedDistributions.GAUSS
+        assert fp_service.extractParameters(target_sequence_default) == ['0', '1']
+
+        line_no_name = 'variableOne <- $gamma(0,1)$;'
+        target_sequence_no_name = self.extractTargetSequenceFromLine(line_no_name)
+        assert fp_service.extractCoefficientName(target_sequence_no_name) == 'coefficient1'
+        assert fp_service.extractDistributionName(target_sequence_no_name) == SupportedDistributions.GAMMA
+        assert fp_service.extractParameters(target_sequence_no_name) == ['0', '1']
+
+        line_quick = 'variableOne <- $0,name=varOne$;'
+        target_sequence_quick = self.extractTargetSequenceFromLine(line_quick)
+        assert fp_service.extractCoefficientName(target_sequence_quick) == 'varOne'
+        assert fp_service.extractDistributionName(target_sequence_quick) == SupportedDistributions.GAUSS
+        assert fp_service.extractParameters(target_sequence_quick) == ['0']
+
+        line_quick_no_name = 'variableOne <- $0$'
+        target_sequence_quick_no_name = self.extractTargetSequenceFromLine(line_quick_no_name)
+        assert fp_service.extractCoefficientName(target_sequence_no_name) == 'coefficient2'
+        assert fp_service.extractDistributionName(target_sequence_quick_no_name) == SupportedDistributions.GAUSS
+        assert fp_service.extractParameters(target_sequence_quick_no_name) == ['0']
+
+        bool_default = '$boolean(0.5), name= B_AKT2$,'
+        target_sequence_bool_default = self.extractTargetSequenceFromLine(bool_default)
+        assert fp_service.extractCoefficientName(target_sequence_bool_default) == 'B_AKT2'
+        assert fp_service.extractDistributionName(target_sequence_bool_default) == SupportedDistributions.BOOLEAN
+        assert fp_service.extractParameters(target_sequence_bool_default) == ['0.5']
+
+        mutate_default = '$mutate(CTNNB1, 0, .304), name= M_CTNNB1$'
+        target_sequence_mutate_default = self.extractTargetSequenceFromLine(mutate_default)
+        assert fp_service.extractCoefficientName(target_sequence_mutate_default) == 'M_CTNNB1'
+        assert fp_service.extractDistributionName(target_sequence_mutate_default) == SupportedDistributions.MUTATE
+        assert fp_service.extractParameters(target_sequence_mutate_default) == ['CTNNB1', '0', '.304']
+
+    def testDollarSignSyntaxValuesCorrectlyExtractedForMATLAB(self):
+        data_file = self.setTargetFile('SampleDataFiles', 'WNT_ERK_crosstalk.m')
+        self.setupFileProcessingService(data_file, SupportedFileTypes.MATLAB)
+        fp_service = self.file_processing_service
+
+        line_default = 'variableOne=$gauss(0,1),name=varOne$;'
+        target_sequence_default = self.extractTargetSequenceFromLine(line_default)
+        assert fp_service.extractCoefficientName(target_sequence_default) == 'varOne'
+        assert fp_service.extractDistributionName(target_sequence_default) == SupportedDistributions.GAUSS
+        assert fp_service.extractParameters(target_sequence_default) == ['0', '1']
+
+        line_no_name = 'variableOne=$gamma(0,1)$;'
+        target_sequence_no_name = self.extractTargetSequenceFromLine(line_no_name)
+        assert fp_service.extractCoefficientName(target_sequence_no_name) == 'coefficient1'
+        assert fp_service.extractDistributionName(target_sequence_no_name) == SupportedDistributions.GAMMA
+        assert fp_service.extractParameters(target_sequence_no_name) == ['0', '1']
+
+        line_quick = 'variableOne=$0,name=varOne$;'
+        target_sequence_quick = self.extractTargetSequenceFromLine(line_quick)
+        assert fp_service.extractCoefficientName(target_sequence_quick) == 'varOne'
+        assert fp_service.extractDistributionName(target_sequence_quick) == SupportedDistributions.GAUSS
+        assert fp_service.extractParameters(target_sequence_quick) == ['0']
+
+        line_quick_no_name = 'variableOne=$0$'
+        target_sequence_quick_no_name = self.extractTargetSequenceFromLine(line_quick_no_name)
+        assert fp_service.extractCoefficientName(target_sequence_no_name) == 'coefficient2'
+        assert fp_service.extractDistributionName(target_sequence_quick_no_name) == SupportedDistributions.GAUSS
+        assert fp_service.extractParameters(target_sequence_quick_no_name) == ['0']
+
+    def extractTargetSequenceFromLine(self, line):
+        search_result = self.file_processing_service.IDENTIFIER_REGEX.search(line)
+        assert search_result is not None
+        return line[(search_result.regs[0][0] + 1):(search_result.regs[0][1] - 1)]
